+++ htof/test/test_parse.py
@@ -3,7 +3,10 @@ import numpy as np
 import pytest
 import mock
 import os
+import tempfile
+from ast import literal_eval
 
+from astropy.table import Table
 from htof.parse import HipparcosOriginalData, HipparcosRereductionData,\
     GaiaData, IntermediateDataParser, GaiaDR2
 from htof.parse import calculate_covariance_matrices, fractional_year_epoch_to_jd, _match_filename_to_star_id
@@ -144,9 +147,9 @@ def test_trim_gaia_data():
     parser = GaiaData()
     datemin, datemax = 3, 5
     epochs = pd.DataFrame(data=[datemin - 1, datemin, datemax, datemax + 1], index=[3, 4, 5, 6])
-    epochs, other = parser.trim_data(epochs, datemin, datemax, other_data=[epochs])
-    assert np.allclose(epochs.values.flatten(), [datemin, datemax])
-    assert np.allclose(other.values.flatten(), [datemin, datemax])
+    data = pd.DataFrame(data=[datemin - 1, datemin, datemax, datemax + 1], index=[3, 4, 5, 6])
+    data = parser.trim_data(data, epochs, datemin, datemax)
+    assert np.allclose(data.values.flatten(), [datemin, datemax])
 
 
 @mock.patch('htof.parse.calculate_covariance_matrices', return_value=np.array([np.ones((2, 2))]))
@@ -181,6 +184,35 @@ class TestParseGaiaData:
         assert np.isclose(data.scan_angle.iloc[0], -0.3066803677989655)
         assert np.isclose(data._epoch.iloc[67], 2458426.7784441216)
         assert np.isclose(data.scan_angle.iloc[67], 2.821818345385301)
+
+
+def test_write_with_missing_info():
+    data = IntermediateDataParser(scan_angle=np.arange(3), epoch=pd.DataFrame(np.arange(1991, 1994)),
+                                  residuals=np.arange(2, 5),
+                                  inverse_covariance_matrix=None,
+                                  along_scan_errs=None)
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        data.write(os.path.join(tmp_dir, 'out.csv'))
+        t = Table.read(os.path.join(tmp_dir, 'out.csv'))
+        assert np.allclose(t['residuals'], data.residuals)
+        assert np.allclose(t['julian_day_epoch'], data.julian_day_epoch())
+        assert np.allclose(t['scan_angle'], data.scan_angle)
+
+
+def test_write():
+    data = IntermediateDataParser(scan_angle=np.arange(3), epoch=pd.DataFrame(np.arange(1991, 1994)),
+                                  residuals=np.arange(2, 5),
+                                  inverse_covariance_matrix=np.array([[1, 2], [3, 4]]) * np.ones((3, 2, 2)),
+                                  along_scan_errs=np.arange(3, 6))
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        data.write(os.path.join(tmp_dir, 'out.csv'))
+        t = Table.read(os.path.join(tmp_dir, 'out.csv'))
+        assert np.allclose(t['residuals'], data.residuals)
+        assert np.allclose(t['julian_day_epoch'], data.julian_day_epoch())
+        assert np.allclose(t['scan_angle'], data.scan_angle)
+        assert np.allclose(t['along_scan_errs'], data.along_scan_errs)
+        icovs = [np.array(literal_eval(icov)) for icov in t['icov']]
+        assert np.allclose(icovs, data.inverse_covariance_matrix)
 
 
 def test_calculating_covariance_matrices():